added readme and more tests

+++ python_web_final_project/companies_app/views.py
@@ -7,11 +7,11 @@ from django.views import generic as views
 from python_web_final_project.accounts_app.forms import RegisterCompanyForm
 from python_web_final_project.accounts_app.views import UserRegisterBaseView
 from python_web_final_project.common_app.models import Job, JobApplication
-from python_web_final_project.companies_app.forms import CompanyEditProfileForm, CompanyAddEditForm
+from python_web_final_project.companies_app.forms import CompanyEditProfileForm, CompanyAddEditJobForm
 from python_web_final_project.companies_app.models import CompanyProfile
 from python_web_final_project.helpers.mixins.custom_user_passes_test_mixins import UserIsCompanyTestMixin, \
     user_is_company
-from python_web_final_project.helpers.mixins.view_mixins import CompanyAddEditJobMixin
+from python_web_final_project.helpers.mixins.view_mixins import CompanyAddEditJobViewMixin
 
 
 class CompanyEditProfileView(LoginRequiredMixin, UserIsCompanyTestMixin, views.UpdateView):
@@ -26,7 +26,7 @@ class CompanyEditProfileView(LoginRequiredMixin, UserIsCompanyTestMixin, views.U
         return get_object_or_404(CompanyProfile, pk=self.request.user.pk)
 
 
-class CompanyAddJobView(LoginRequiredMixin, CompanyAddEditJobMixin, UserIsCompanyTestMixin, views.CreateView):
+class CompanyAddJobView(LoginRequiredMixin, CompanyAddEditJobViewMixin, UserIsCompanyTestMixin, views.CreateView):
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         context['form_action'] = 'Add Job'
@@ -37,7 +37,7 @@ class CompanyAddJobView(LoginRequiredMixin, CompanyAddEditJobMixin, UserIsCompan
         return success_url
 
 
-class CompanyEditJobView(LoginRequiredMixin, CompanyAddEditJobMixin, UserIsCompanyTestMixin, views.UpdateView):
+class CompanyEditJobView(LoginRequiredMixin, CompanyAddEditJobViewMixin, UserIsCompanyTestMixin, views.UpdateView):
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -64,7 +64,8 @@ class CompanyJobApplicationsView(LoginRequiredMixin, UserIsCompanyTestMixin, vie
 @user_passes_test(user_is_company, login_url=reverse_lazy('index'))
 def delete_job(request, pk):
     job = get_object_or_404(Job, pk=pk)
-    if request.method == 'POST':
+    if request.method == 'POST' and job.company == request.user:
         job.delete()
         return redirect(reverse_lazy('company jobs', kwargs={'pk': request.user.pk}))
     return redirect(reverse('job details', kwargs={'pk': job.pk}))
+
